Move graph code from activity to fragment
